The implicit cast looked a bit scary (at least hard to understand). This did not solve the crash I am seeing, though.
[git-p4: depot-paths = "//depot/qt/main/": change = 45989]

+++ src/widgets/qaction.cpp
@@ -571,7 +571,7 @@ void QAction::setAccel( const QKeySequence& key )
     d->accel = 0;
 #endif
 
-    if ( !key ) {
+    if ( !(int)key ) {
 	d->update();
 	return;
     }
